requirements.txt
python-dotenv
discord
discord-py-slash-command
twitchio